Refactor api to return a palette object.
This removes the `format` option when creating a palette in favour of calling `.format()` on a returned palette object.
The main benefit of this approach is that the return type from .format() is guaranteed and will vary based on the input argument.
It also allows the creation of a palette once before using it in several formats later.